Link patient IDs to geocoded variables for Table 9
This file is separate from the rest of the tables because it need to be run against an identified database. The end result is the Status_Variables, which contains the geocoded variables for each patient, provided that patient has a tract ID associated with it. 

NOTE: About 70% of patients came back with no tract ID, for reasons not yet determined.